// src/Moryx.Operators.Web/src/app/api/models/assignable-operator.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { IOperatorAssignable } from '../models/i-operator-assignable';
export interface AssignableOperator {
  assignedResources?: Array<IOperatorAssignable> | null;
  firstName?: string | null;
  identifier?: string | null;
  lastName?: string | null;
  pseudonym?: string | null;
  signedIn?: boolean;
}